+++ lib/extensions/email_confirmation/phoenix/router.ex
@@ -2,9 +2,11 @@ defmodule PowEmailConfirmation.Phoenix.Router do
   @moduledoc false
   use Pow.Extension.Phoenix.Router.Base
 
+  alias Pow.Phoenix.Router
+
   defmacro routes(_config) do
     quote location: :keep do
-      resources "/confirm-email", ConfirmationController, only: [:show]
+      Router.pow_resources "/confirm-email", ConfirmationController, only: [:show]
     end
   end
 end